\name{DemoData}

\alias{gbpusd}
\alias{spx}

\docType{data}

\title{Demo Data}

\description{

  Price datasets to demonstrate functionality of the \code{tradesys} package.

}

\details{

  \code{gbpusd} is a vector of the daily cable rate since 05 January
  1971, as reported in the in the FRBNY H10 release.
  
  \code{spx} is a data frame of the daily OHLC series for the S&P 500
  since 03 January 1950.

}

\references{
  \tabular{l}{
    \url{http://www.federalreserve.gov/releases/h10/hist/} \cr
    \url{http://finance.yahoo.com/}
  }
}

\author{Robert Sams \email{robert@sanctumfi.com}}

\keyword{math}
