\name{RollingFunctions}

\alias{MA}
\alias{HIGH}
\alias{LOW}

\title{Rolling Functions}

\description{
  Optimized rolling calculations for use within trading system signal statements.
}

\usage{
MA(x, width)
HIGH(x, width)
LOW(x, width)
}

\arguments{
  \item{x}{numeric vector.}
  \item{width}{numeric.}
}

\details{
  NEED TO DOCUMENT

  Add: LAG, TR, ATR, EWMA, etc.
}

\author{Robert Sams \email{robert@sanctumfi.com}}

\keyword{math}